fixes reboot on RHEL8 (#67)
fixes https://github.com/linux-system-roles/sap-hana-preconfigure/issues/67

The command needs-restarting is part of the package yum-utils, which is not in the list of required packages (i.e.@base or @server) This command is used to check if a reboot of the server is required. Hence yum-utils is added to the list of required packages.

This command is used in the module "check if reboot is required (7.3+)"
Currently this module is only executed on RHEL7, so the when statement is changed accordingly.
If the command is not installed this module fails.

If the module named "check if reboot is required (7.3+)" is skipped the according variable "sap_hana_preconfigure_register_reboot_required" is not set.
If the default value of "sap_hana_preconfigure_reboot_after_update" is changed to true and sap_hana_preconfigure_register_reboot_required is not set the playbook will fail with unset variable.